add global setup to playwright
