+++ playwright.config.ts
@@ -31,7 +31,7 @@ export default defineConfig({
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
   },
-
+  globalSetup: require.resolve('./tests/global-setup'),
   /* Configure projects for major browsers */
   projects: [
     {
