+++ tests/disconnected/homepage/desktop.spec.ts
@@ -1,16 +1,5 @@
 import { test } from "@playwright/test";
 
-const validTenants = ["optimism", "ens", "etherfi", "lyra"];
-const tenant: string = process.env.NEXT_PUBLIC_AGORA_INSTANCE_NAME || "";
-
-if (!validTenants.includes(tenant)) {
-    throw new Error(
-        `NEXT_PUBLIC_AGORA_INSTANCE_NAME must be one of ${validTenants.join(
-            ", "
-        )}, but got '${tenant}'`
-    );
-}
-
 test.describe('homepage desktop', () => {
     test('connect wallet button', async ({ page }) => {
         await page.goto('/');
