+++ tests/disconnected/homepage/metadata.spec.ts
@@ -1,20 +1,5 @@
 import { test, expect } from "@playwright/test";
 
-// TODO: frh -> move all this into a fixture
-// TODO: frh -> typescript
-// TODO: frh -> tests are failing on other browsers and connect wallet
-const validTenants = ["optimism", "ens", "etherfi", "lyra"];
-const tenant: string = process.env.NEXT_PUBLIC_AGORA_INSTANCE_NAME || "";
-
-if (!validTenants.includes(tenant)) {
-    throw new Error(
-        `NEXT_PUBLIC_AGORA_INSTANCE_NAME must be one of ${validTenants.join(
-            ", "
-        )}, but got '${tenant}'`
-    );
-}
-
-
 test('homepage metadata', async ({ page }) => {
     await page.goto('/');
